backend/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.3
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.15.1
e3fp==1.2.5
et-xmlfile==1.1.0
idna==3.7
isort==5.13.2
joblib==1.4.2
mmh3==4.1.0
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.2
pillow==10.3.0
python-dateutil==2.9.0.post0
pytz==2024.1
rdkit==2023.9.6
requests==2.32.1
scikit-learn==1.4.2
scipy==1.13.0
sdaxen-python-utilities==0.1.5
six==1.16.0
smart-open==7.0.4
sqlparse==0.4.4
threadpoolctl==3.5.0
tzdata==2024.1
urllib3==2.2.1
wrapt==1.16.0
